pyproject.toml
[tool.poetry]
name = "rest2"
version = "0.1.0"
description = "REST API"
authors = ["villanelle3 <106897289+villanelle3@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.3"
factory-boy = "^3.3.0"
django = "^5.0"
djangorestframework = "^3.14.0"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
